fix off_t abuse in pagebuf don't use off_t for pagebuf variables

+++ fs/xfs/pagebuf/page_buf.h
@@ -340,11 +340,13 @@ extern int pagebuf_iorequest(		/* start real I/O		*/
 extern int pagebuf_iowait(		/* wait for buffer I/O done	*/
 		page_buf_t *);		/* buffer to wait on		*/
 
-extern caddr_t	pagebuf_offset(page_buf_t *, off_t);
+extern caddr_t	pagebuf_offset(
+		page_buf_t *,
+		unsigned int);
 
 extern void pagebuf_iomove(		/* move data in/out of pagebuf	*/
 		page_buf_t *,		/* buffer to manipulate		*/
-		off_t,			/* starting buffer offset	*/
+		loff_t,			/* starting buffer offset	*/
 		size_t,			/* length in buffer		*/
 		caddr_t,		/* data pointer			*/
 		page_buf_rw_t);		/* direction			*/
